fixed nothing

+++ libraries/graphics/renderer.rs
@@ -291,12 +291,12 @@ impl Renderer {
                 gl::BlitFramebuffer(
                     0,
                     0,
-                    self.get_window_size().0 as i32 * 2,
-                    self.get_window_size().1 as i32 * 2,
+                    (self.get_window_size().0 * 2.0) as i32,
+                    (self.get_window_size().1 * 2.0) as i32,
                     0,
                     0,
-                    self.get_window_size().0 as i32 * 2,
-                    self.get_window_size().1 as i32 * 2,
+                    (self.get_window_size().0 * 2.0) as i32,
+                    (self.get_window_size().1 * 2.0) as i32,
                     gl::COLOR_BUFFER_BIT,
                     gl::NEAREST,
                 );
@@ -310,8 +310,8 @@ impl Renderer {
                     (self.get_window_size().1 * 2.0) as i32,
                     0,
                     0,
-                    self.get_window_size().0 as i32 * 2,
-                    self.get_window_size().1 as i32 * 2,
+                    (self.get_window_size().0 * 2.0) as i32,
+                    (self.get_window_size().1 * 2.0) as i32,
                     gl::COLOR_BUFFER_BIT,
                     gl::NEAREST,
                 );
@@ -325,8 +325,8 @@ impl Renderer {
                     (self.get_window_size().1 as f32 * 2.0) as i32,
                     0,
                     0,
-                    self.get_window_size().0 as i32 * 2,
-                    self.get_window_size().1 as i32 * 2,
+                    (self.get_window_size().0 as f32 * 2.0) as i32,
+                    (self.get_window_size().1 as f32 * 2.0) as i32,
                     gl::COLOR_BUFFER_BIT,
                     gl::NEAREST,
                 );
